Add K-Link example deploy configuration and related documentation (#2)
This pull request adds an example docker-compose.yml file and the necessary documentation to perform a local deployment using the example configuration. In addition add reverse proxy configuration.

At the K-Link explore site level, this pull request, change the location of the source to the /source folder and enhance the local NGINX configuration to proxy the K-Search and the K-Registry. For the K-Search add also the necessary CORS headers.

Commits:
* Move html and static content to the source directory
* Improve docker image building and NGINX default configuration
* Fix docker image name and add reference to source folder
* Support K-Registry proxy
* Add example docker-compose configuration
* Add installation and deploy documentation. Improve readme